Refactor local data into program directory
