+++ dpc/src/testnet1/mod.rs
@@ -108,27 +108,6 @@ pub trait Testnet1Components: DPCComponents {
 
 ///////////////////////////////////////////////////////////////////////////////
 
-/// Stores local data required to produce program proofs.
-pub struct LocalData<C: Testnet1Components> {
-    pub system_parameters: SystemParameters<C>,
-
-    // Old records and serial numbers
-    pub old_records: Vec<Record<C>>,
-    pub old_serial_numbers: Vec<<C::AccountSignature as SignatureScheme>::PublicKey>,
-
-    // New records
-    pub new_records: Vec<Record<C>>,
-
-    // Commitment to the above information.
-    pub local_data_merkle_tree: CommitmentMerkleTree<C::LocalDataCommitment, C::LocalDataCRH>,
-    pub local_data_commitment_randomizers: Vec<<C::LocalDataCommitment as CommitmentScheme>::Randomness>,
-
-    pub memorandum: <Transaction<C> as TransactionScheme>::Memorandum,
-    pub network_id: u8,
-}
-
-///////////////////////////////////////////////////////////////////////////////
-
 pub struct DPC<C: Testnet1Components> {
     _components: PhantomData<C>,
 }
