+++ dpc/src/testnet1/program/mod.rs
@@ -14,6 +14,9 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
+pub mod local_data;
+pub use local_data::*;
+
 pub mod noop_program;
 pub use noop_program::*;
